PS2 Mouse: Moved mouse_ps2 driver to module
