+++ app/module/drivers/Kconfig
@@ -5,3 +5,4 @@ rsource "gpio/Kconfig"
 rsource "kscan/Kconfig"
 rsource "sensor/Kconfig"
 rsource "display/Kconfig"
+rsource "input/Kconfig"
